feat(showList): Create Show model
