+++ AmazingApp/AmazingApp/ShowsList/Service/ShowRepository.swift
@@ -38,39 +38,7 @@ enum ShowEndpoint: Endpoint {
     
 }
 
-struct Show: Decodable {
-    
-    let id: Int?
-    let name: String?
-    let genres: [String]?
-    let schedule: Schedule?
-    let summary: String?
-    let image: ImagePath?
-    
-    struct Schedule: Decodable {
-        let time: String?
-        let days: [Day]
-        
-        enum Day: String, Decodable {
-            case monday = "Monday",
-                tuesday = "Tuesday",
-                wednesday = "Wednesday",
-                thursday = "Thursday",
-                friday = "Friday",
-                saturday = "Saturday",
-                sunday = "Sunday"
-        }
-    }
-    
-    struct ImagePath: Decodable {
-        let medium: String?
-    }
-    
-    struct Rating: Decodable {
-        let average: Double?
-    }
-    
-}
+
 
 final class ShowRepository: ShowRepositoryType {
